.gitignore is now working

+++ administration/views.py
@@ -5,12 +5,12 @@ from propose_join.models import ExistingClub,ProposedClub
 from propose_join.forms import club_logo
 from django.contrib.auth import get_user_model
 from .models import Onpollclub
+from django.contrib.auth.models import User
+from registration.models import  Profile
 from datetime import datetime
 from django.utils import timezone
 
 from dateutil.relativedelta import relativedelta
-
-User = get_user_model()
 
 def index(request):
     qs = ProposedClub.objects.all()
@@ -18,15 +18,22 @@ def index(request):
 
 def result(request):
     ab = ProposedClub.objects.all()
+    cs=None
     for a in ab:
-        a.delete()
-        break;
+        if a.club_name == request.POST['name']:
+            print(a)
+            cs = a.name
+            a.delete()
+            break;
+
+    mail=User.objects.get(username=cs).email
+
     fullname1 = request.POST['message']
     send_mail(
         'Your club is rejected',
         fullname1,
         'iiits2021@gmail.com',
-        ['saisreeramputta156@gmail.com'],
+        [mail],
         fail_silently=False,
     )
     qs = ProposedClub.objects.all()
@@ -35,15 +42,20 @@ def result(request):
 # Create your views here.
 def result1(request):
     ab = ProposedClub.objects.all()
+    cs=None
     for a in ab:
-        Onpollclub.objects.create(club_name=a.club_name,admin=a.name,club_info=a.club_info,club_logo=a.club_logo)
-        a.delete()
-        break;
+        if a.club_name == request.POST['name']:
+            print(a)
+            Onpollclub.objects.create(club_name=a.club_name,admin=a.name,club_info=a.club_info,club_logo=a.club_logo)
+            cs = a.name
+            a.delete()
+            break;
+    mail = User.objects.get(username=cs).email
     send_mail(
         'Your club is accepted',
-        'fullname1',
+        'Hii,              Tell all your friends to participate in polls                                 thanks ',
         'iiits2021@gmail.com',
-        ['saisreeramputta156@gmail.com'],
+        [mail],
         fail_silently=False,
     )
     qs = ProposedClub.objects.all()
